refactor(log): move log repository to domain/log/repositories
Centralize log repository interface in domain/log/repositories for better organization and maintainability. Update all imports to reflect the new location.
