+++ src/domain/log/use-cases/__test__/create-log.test.ts
@@ -1,5 +1,5 @@
 import { CreateLog } from "../create-log";
-import { LogRepository } from "@/domain/repositories/log.repository";
+import { LogRepository } from "@/domain/log/repositories/log.repository";
 import { CreateLogDto } from "@/domain/log/dtos/create-log.dto";
 import { LogEntity } from "@/domain/entities/log.entity";
