+++ src/domain/log/use-cases/__test__/delete-old-logs.test.ts
@@ -1,5 +1,5 @@
 import { DeleteOldLogs } from "@/domain/log/use-cases/delete-old-logs";
-import { LogRepository } from "@/domain/repositories/log.repository";
+import { LogRepository } from "@/domain/log/repositories/log.repository";
 
 describe("DeleteOldLogs UseCase", () => {
   it("delegates to repository.deleteOldLogs and returns count", async () => {
