+++ src/domain/log/use-cases/delete-old-logs.ts
@@ -1,4 +1,4 @@
-import { LogRepository } from "@/domain/repositories/log.repository";
+import { LogRepository } from "@/domain/log/repositories/log.repository";
 
 interface IDeleteOldLogsUseCase {
   execute(olderThanDays: number): Promise<number>;
